Fold the activity grid into the reading summary
Review feedback on #706: the grid is more of the same statistics, not a
section of its own, so it loses its heading and sits under the numbers inside
`ReadingStatsSection`. That section already owns the statistics query and its
failure reporting, so the grid stops making a second call of its own and
takes the activity as a prop -- one query, one place failures are handled.

It also loses its empty state. `ReadingStatsSection` stands down entirely for
a book nobody has opened, and the grid now stands down with it rather than
reporting an absence beside a summary that is already absent.

Every square gains `role="img"` and a label saying what its day came to, via
`renderBlock`. `frontend/claude.md` asks for queries by accessible role and
name, and the tests were reaching into the calendar's own `data-date` and
`data-level` attributes instead. The label is worth more than the test
though: a tooltip needs a pointer to hover, so on a phone this is the only
way to a day's number at all.

Comments trimmed as asked -- the block weighing block-size options against
each other, and the note explaining the weekday-label rotation, which the
code no longer does. The week-start comment now says what the call actually
delivers: Firefox exposes no `Intl.Locale` week info, so Luxon falls back to
Monday there whatever the reader's locale.
