+++ frontend/src/pages/BookPage/ReadingStatistics/ReadingStatsSection.tsx
@@ -4,6 +4,7 @@ import { countLabel } from '@/utils/counts.ts';
 import { browserTimeZone, formatDate, formatSeconds } from '@/utils/date.ts';
 import { Box, LinearProgress, Typography } from '@mui/material';
 import { useEffect } from 'react';
+import { ReadingActivityGrid } from './ReadingActivityGrid.tsx';
 
 interface ReadingStatsSectionProps {
   bookId: number;
@@ -110,6 +111,9 @@ export const ReadingStatsSection = ({ bookId }: ReadingStatsSectionProps) => {
     <Box sx={{ mb: 4 }}>
       {progress != null && <ReadingProgress percent={progress} />}
       <StatsGrid stats={stats} />
+      {/* The grid is more of the same summary, not a section of its own, so it
+          sits under the numbers rather than beside them with a heading. */}
+      {data.activity && <ReadingActivityGrid activity={data.activity} />}
     </Box>
   );
 };
